backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
playwright==1.42.0
beautifulsoup4==4.12.3
pandas==2.2.0
openpyxl==3.1.2
fpdf2==2.7.6
loguru==0.7.2
apscheduler==3.10.4
slowapi==0.1.8
python-multipart==0.0.9
Pillow==10.2.0
httpx==0.27.0
